Collider (1)
- 컴포넌트 구조. 콜라이더 클래스 정의 중. 오브젝트에서 콜라이더 포인터를 멤버로 두고 nullptr로 두면 안 쓰는거고 안 두면 쓰는거.
- 콜라이더가 부착된 오브젝트를 따라다니는 것을 finalupdate()에서 구현. 내부적인 오브젝트 이동 같은 것들 다 끝난 후에 final update에서 콜라이더 움직임이 실행됨.
- SelectGDI 클래스 만듦. 렌더할 때 pen/brush 컨트롤 할 수 있도록 Core 코드 추가하고, SelectGDI 객체 생성하면 원하는 펜 브러쉬를 생성자에서 설정하고 호출된 함수 종료되면서 소멸자 호출되어서 소멸자에서 다시 defaultpen brush로 돌려놓는 작업 실행.

+++ Project2/CPathMgr.cpp
@@ -26,7 +26,7 @@ void CPathMgr::init()
 	
 	// system 쪽 라이브러리 참조해서 상위폴더 함수 사용할 수 있지만 문자열로 다뤄보겠다
 	// 맨 뒤부터 차례대로 보다가 \\가 나오면 걔를 \0 널문자로 변경. 그 뒤를 다 제거할 필요 없음.
-	int iLen = wcslen(m_szContentPath);
+	int iLen = (int)wcslen(m_szContentPath);
 	for (int i = iLen-1; i >= 0; --i)
 	{
 		if (m_szContentPath[i] == '\\')
